Changes to HV direct solicited
Only applies this when BOTH the target audience is HV and the product is direct solicited. Further to conversation with Ado today. So this will put less in HV direct solicited but more in (probably) HV stewardship and more in 'other warm fundraising' or similar, including because some of the HV upgrade appeals are combined with cash appeals and have cash products(!) e.g. 14/HAC